sample_feed/requirements.txt
Django>=3.2,<4.0
gunicorn>=20.0,<21.0
gevent
djangorestframework>=3.12,<4.0
redis
pymysql
requests
icecream